requirements.txt
stix2
taxii2-client
python-dotenv
